+++ languages.go
@@ -3,6 +3,7 @@ package suzu
 import (
 	"fmt"
 
+	"github.com/aws/aws-sdk-go-v2/service/transcribestreaming/types"
 	"github.com/aws/aws-sdk-go/service/transcribestreamingservice"
 )
 
@@ -22,10 +23,18 @@ func GetLanguageCode(serviceType, lang string, f func(string) (string, error)) (
 	}
 
 	switch serviceType {
-	case "aws":
+	case "awsv1":
 		for _, languageCode := range transcribestreamingservice.LanguageCode_Values() {
 			if languageCode == lang {
 				return languageCode, nil
+			}
+		}
+		return "", fmt.Errorf("%w: %s", ErrUnsupportedLanguageCode, lang)
+	case "aws", "awsv2":
+		lc := new(types.LanguageCode)
+		for _, languageCode := range lc.Values() {
+			if languageCode == types.LanguageCode(lang) {
+				return lang, nil
 			}
 		}
 		return "", fmt.Errorf("%w: %s", ErrUnsupportedLanguageCode, lang)